+++ tests/unit/api/versionEndpoints.spec.ts
@@ -7,6 +7,8 @@ import ExaminationRegulationManagement from "@/api/ExaminationRegulationManageme
 import ConfigurationManagement from "@/api/ConfigurationManagement";
 import AdmissionManagement from "@/api/AdmissionManagement";
 import ReportManagement from "@/api/ReportManagement";
+import OperationManagement from "@/api/OperationManagement";
+import GroupManagement from "@/api/GroupManagement";
 
 jest.useFakeTimers();
 jest.setTimeout(30000);
@@ -80,6 +82,26 @@ test("admission service version", async () => {
     expect(version).not.toEqual("unavailable");
 
     const hyperledgerVersion = await AdmissionManagement.getHyperledgerVersion();
+    expect(hyperledgerVersion.chaincodeVersion).not.toBe(undefined);
+    expect(hyperledgerVersion.hlfApiVersion).not.toBe(undefined);
+});
+
+test("operation service version", async () => {
+    const version = await OperationManagement.getVersion();
+
+    expect(version).not.toEqual("unavailable");
+
+    const hyperledgerVersion = await OperationManagement.getHyperledgerVersion();
+    expect(hyperledgerVersion.chaincodeVersion).not.toBe(undefined);
+    expect(hyperledgerVersion.hlfApiVersion).not.toBe(undefined);
+});
+
+test("group service version", async () => {
+    const version = await GroupManagement.getVersion();
+
+    expect(version).not.toEqual("unavailable");
+
+    const hyperledgerVersion = await GroupManagement.getHyperledgerVersion();
     expect(hyperledgerVersion.chaincodeVersion).not.toBe(undefined);
     expect(hyperledgerVersion.hlfApiVersion).not.toBe(undefined);
 });